Move get_host_* to spack.spec
These functions need both spack.spec and spack.platform
